using instead customOp, ddirect_register_custom_op, adding it to benchmarking and resturcure the code
